Tback/requirements.txt
Django==5.2.5
djangorestframework==3.16.1
django-cors-headers==4.7.0
django-filter==25.1
stripe==12.5.0b2
python-dotenv==1.0.0
requests==2.31.0